# optimize rows_collection_parser (cf. Roo::Utils)