… lib/saxlsx/column_name_generator.rb
… spec/column_name_generator_spec.rb