Using Roo's implementation of rows_collection_parser gives a big performance boost 

### Before

```
Shared Strings

Rehearsal --------------------------------------------------------
creek                  0.740139   0.038133   0.778272 (  0.779112)
dullard                0.556777   0.008871   0.565648 (  0.566204)
oxcelix                0.701892   0.024388   0.726280 (  0.726728)
roo                    0.902212   0.028503   0.930715 (  0.933237)
rubyXL                 2.159253   0.061215   2.220468 (  2.226967)
saxlsx                 0.463630   0.021618   0.485248 (  0.485915)
simple_xlsx_reader     0.729623   0.026273   0.755896 (  0.757106)
----------------------------------------------- total: 6.462527sec

                           user     system      total        real
creek                  0.677348   0.026071   0.703419 (  0.703911)
dullard                0.520397   0.011546   0.531943 (  0.532647)
oxcelix                0.665139   0.023925   0.689064 (  0.690179)
roo                    0.767174   0.018318   0.785492 (  0.787596)
rubyXL                 1.924527   0.024093   1.948620 (  1.950305)
saxlsx                 0.338003   0.009786   0.347789 (  0.348505)
simple_xlsx_reader     0.641956   0.019137   0.661093 (  0.661816)

Inline Strings

Rehearsal --------------------------------------------------------
creek                  0.892396   0.095330   0.987726 (  0.988973)
dullard                0.747026   0.100259   0.847285 (  0.847985)
oxcelix              ERROR
  0.030298   0.023939   0.054237 (  0.054569)
roo                    1.243016   0.051569   1.294585 (  1.300220)
rubyXL                 2.556644   0.101820   2.658464 (  2.661338)
saxlsx                 0.737219   0.087095   0.824314 (  0.825358)
simple_xlsx_reader     0.844790   0.087676   0.932466 (  0.933846)
----------------------------------------------- total: 7.599077sec

                           user     system      total        real
creek                  0.956009   0.203042   1.159051 (  1.167953)
dullard                0.643613   0.079597   0.723210 (  0.724083)
oxcelix              ERROR
  0.025964   0.026075   0.052039 (  0.052367)
roo                    1.133606   0.040469   1.174075 (  1.178371)
rubyXL                 2.200222   0.071803   2.272025 (  2.273808)
saxlsx                 0.559796   0.077359   0.637155 (  0.637769)
simple_xlsx_reader     0.806133   0.085169   0.891302 (  0.892003)
```

### After
```
Shared Strings

Rehearsal --------------------------------------------------------
creek                  0.740222   0.042622   0.782844 (  0.784274)
dullard                0.553352   0.007750   0.561102 (  0.561806)
oxcelix                0.690075   0.026774   0.716849 (  0.718120)
roo                    0.901606   0.024028   0.925634 (  0.927313)
rubyXL                 2.109617   0.060372   2.169989 (  2.173096)
saxlsx                 0.292785   0.015283   0.308068 (  0.308619)
simple_xlsx_reader     0.753668   0.031550   0.785218 (  0.785726)
----------------------------------------------- total: 6.249704sec

                           user     system      total        real
creek                  0.674811   0.025639   0.700450 (  0.701539)
dullard                0.525192   0.011903   0.537095 (  0.537667)
oxcelix                0.721473   0.032590   0.754063 (  0.785717)
roo                    0.865110   0.024254   0.889364 (  0.898208)
rubyXL                 2.209751   0.042147   2.251898 (  2.272176)
saxlsx                 0.248144   0.005913   0.254057 (  0.255748)
simple_xlsx_reader     0.741901   0.012591   0.754492 (  0.761650)

Inline Strings

Rehearsal --------------------------------------------------------
creek                  1.008018   0.097117   1.105135 (  1.115229)
dullard                0.843527   0.082316   0.925843 (  0.937090)
oxcelix              ERROR
  0.033292   0.016246   0.049538 (  0.052278)
roo                    1.207119   0.054048   1.261167 (  1.263271)
rubyXL                 2.502212   0.095102   2.597314 (  2.599455)
saxlsx                 0.600547   0.107610   0.708157 (  0.709058)
simple_xlsx_reader     0.845328   0.104901   0.950229 (  0.951402)
----------------------------------------------- total: 7.597383sec

                           user     system      total        real
creek                  0.995564   0.190987   1.186551 (  1.202557)
dullard                0.732562   0.071650   0.804212 (  0.810516)
oxcelix              ERROR
  0.025282   0.022985   0.048267 (  0.048502)
roo                    1.209069   0.046376   1.255445 (  1.264786)
rubyXL                 2.436497   0.106442   2.542939 (  2.565960)
saxlsx                 0.439999   0.133351   0.573350 (  0.574136)
simple_xlsx_reader     0.812258   0.108058   0.920316 (  0.921612)
```